+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 
 2.0.2
     1、ExposedParameterView 修改NiceName
+    2、ParameterNode增加默认值处理 ExposedParameterDefaultValueFunc
+    3、黑板增加排序按钮
+    4、黑板加号，增加排序
+    5、黑板增加SetOrAddParameterValue函数